Found problem with db added to comments for later

+++ app/src/main/java/com/nobos/moneymap/models/Budget.kt
@@ -3,14 +3,14 @@ package com.nobos.moneymap.models
 // Budget.kt
 data class Budget(
     val userId: String = "",
-    val income: Int = 0,
-    val foodExpense: Int = 0,
-    val gasExpense: Int = 0,
-    val entertainmentExpense: Int = 0,
-    val savings: Int = 0,
-    val day: Int = 1,
-    val month: Int = 0,
-    val year: Int = 0,
+    val income: Long = 0,
+    val foodExpense: Long = 0,
+    val gasExpense: Long = 0,
+    val entertainmentExpense: Long = 0,
+    val savings: Long = 0,
+    val day: Long = 1,
+    val month: Long = 0,
+    val year: Long = 0,
     val timestamp: Long = System.currentTimeMillis()
 )
